feat(client): add NoRegister component
